+++ Task_02/Enums/HeroNames.cs
@@ -1,4 +1,4 @@
-﻿namespace Task_02
+﻿namespace Task_02.Enums
 {
     internal static class HeroNames
     {
