+++ Task_02/Printers/Abstract/WinRate/WinRatePrinter.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Task_02.Entities;
 
 namespace Task_02.Printers.Abstract.WinRate
 {
