+++ Task_02/Statistics/Certain/HeroUsagesStatisticProvider.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Task_02.Entities;
 using Task_02.Printers.Certain;
 
 namespace Task_02.Statistics.Certain
