+++ Task_02/Statistics/HeroUsagesStatisticCalculator.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Data.SqlClient;
 using System.Linq;
+using Task_02.Entities;
 
 namespace Task_02.Statistics
 {
@@ -11,12 +12,12 @@ namespace Task_02.Statistics
 
         public List<HeroScoreInfo> FindMostSuccessfulHeros(IEnumerable<HeroScoreInfo> heroScoreInfos)
         {
-            return FindHerosByCondition(heroScoreInfos, MatchScoreInfoCalculator.CalculateWinRate, SortOrder.Descending);
+            return FindHerosByCondition(heroScoreInfos, ScoreInfoCalculator.CalculateWinRate, SortOrder.Descending);
         }
 
         public List<HeroScoreInfo> FindMostUnsuccessfulHeros(IEnumerable<HeroScoreInfo> heroScoreInfos)
         {
-            return FindHerosByCondition(heroScoreInfos, MatchScoreInfoCalculator.CalculateWinRate, SortOrder.Ascending);
+            return FindHerosByCondition(heroScoreInfos, ScoreInfoCalculator.CalculateWinRate, SortOrder.Ascending);
         }
 
         #endregion Find Successful
@@ -25,12 +26,12 @@ namespace Task_02.Statistics
 
         public List<HeroScoreInfo> FindMostFavouriteHeros(IEnumerable<HeroScoreInfo> heroScoreInfos)
         {
-            return FindHerosByCondition(heroScoreInfos, MatchScoreInfoCalculator.CalculateNumberOfPlayedMatches, SortOrder.Descending);
+            return FindHerosByCondition(heroScoreInfos, ScoreInfoCalculator.CalculateNumberOfPlayedMatches, SortOrder.Descending);
         }
 
         public List<HeroScoreInfo> FindMostUnfavouriteHeros(IEnumerable<HeroScoreInfo> heroScoreInfos)
         {
-            return FindHerosByCondition(heroScoreInfos, MatchScoreInfoCalculator.CalculateNumberOfPlayedMatches, SortOrder.Ascending);
+            return FindHerosByCondition(heroScoreInfos, ScoreInfoCalculator.CalculateNumberOfPlayedMatches, SortOrder.Ascending);
         }
 
         #endregion Find Favourite
@@ -39,7 +40,7 @@ namespace Task_02.Statistics
 
         public List<HeroScoreInfo> FindMostWinStreakHeros(IEnumerable<HeroScoreInfo> heroScoreInfos)
         {
-            return FindHerosByCondition(heroScoreInfos, MatchScoreInfoCalculator.CalculateWinStreak, SortOrder.Descending);
+            return FindHerosByCondition(heroScoreInfos, ScoreInfoCalculator.CalculateWinStreak, SortOrder.Descending);
         }
 
         #endregion Find WinStreak
